m/t/launch/cluster: use Health for node liveness in RebootNode
Depending on just heartbeat timestamps is buggy. A node's Health is
calculated on demand, and the calculation is aware of expected heartbeat
intervals configured within the cluster, let's use that, gated by an
additional condition to make sure that the last heartbeat has been
issued after the check has started.

Change-Id: Id5982182a43ae9371b7d25451fedf91df3329218
Reviewed-on: https://review.monogon.dev/c/monogon/+/2882
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>