m/pkg/supervisor: close connections when grpc server exits
When the Listener is closed, Serve will return with an error, but 
already established connections will continue to serve requests. Stop or 
GracefulStop must be called to close these connections.

This bug often caused the metropolis e2e test to fail on my machine with 
the same symptoms as in #276: Node commit always failed with "lost 
leadership". This happened because the nodes were sending requests on a 
connection that was established before the leader was re-elected and the 
grpc listener restarted, and still had the old leadership info.

Change-Id: I797ffa4a40914e5d45e7e4cd15fbb7655e930fa3
Reviewed-on: https://review.monogon.dev/c/monogon/+/2885
Reviewed-by: Serge Bazanski <serge@monogon.tech>
Tested-by: Jenkins CI
Vouch-Run-CI: Serge Bazanski <serge@monogon.tech>