m/n/k/reconciler: refactor resource interface
Replace interface{} with meta.Object, an interface which provides 
accessors for and is implemented by meta.ObjectMeta. List now returns 
the objects themselves instead of their names. This makes the reconciler 
slightly less generic, as it now only supports kubernetes objects.

This is a refactoring in preparation for implementing updates in the 
reconciler. There should be no change in behavior.

Change-Id: I97a4b1c0166a1e6fd0f247ee04e7c44cff570fd7
Reviewed-on: https://review.monogon.dev/c/monogon/+/2891
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI
Vouch-Run-CI: Lorenz Brun <lorenz@monogon.tech>