m/p/erofs: handle metadata alignment correctly
allocateMetadata has an alignment argument but doesn't pass that down
to findBlocks, instead opting to hardcode it to 32.
Pass down the alignment argument and add support for zero alignment to
findBlocks.

This does not cause any functional changes as this function was only
ever called with 32 alignment and once at the zero position with 0
alignment, but the zero position is aligned to every possible
alignment anyways.

Change-Id: Iad44e2a5ae5a99cf9449bba3ec88116e0e009b64
Reviewed-on: https://review.monogon.dev/c/monogon/+/2987
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>
Tested-by: Jenkins CI