metropolis: replace node pubkey with id in cluster directory
The pubkey in the cluster directory was never used in the first place,
and its presence complicates things (exposes the notion of nodes having
a public key and potentially causes reliance on the public key in
untrusted scenarios, eg. when the cluster directory is persisted to
untrusted storage).

Let's just replace that field with a string node ID instead, as that's
also immediately useful in future code (persisting host name mapping
between reboots).

Change-Id: I9b29ee36974ef6edce6076b5df1b8b330fef8bd8
Reviewed-on: https://review.monogon.dev/c/monogon/+/2981
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>