m/n/core/n/hostsfile: persist and load node names to/from Cluster Directory
This change persists node IDs alongside their addresses to the ESP
ClusterDirectory, and loads them up on startup to pre-populate
/etc/hosts.

This is important to bring up clusters from a full cold shutdown, as
these name/address mappings are needed by etcd to connect to other
nodes (etcd connects to other members over metropolis node names, and
expects to be able to resolve them to actual IP addresses).

This didn't affect rolling restarts as other nodes would connect to a
newly started up node. But if all nodes are down, this is needed to
actualy run.

Change-Id: Ifa944bb231909983af2fcb9418a2769e7af65509
Reviewed-on: https://review.monogon.dev/c/monogon/+/2989
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>