m/p/tpm/e/internal: fix validation logic error
The previous logic was wrong: (et < 0x80000000 && et > 0x800000FF) 
always evaluates to false. The overall condition evaluates to true for 
values in the range [0x80000000, 0x800000FF] and thus incorrectly 
rejects them as invalid.

Now, an error is only returned if the value is not in either of the two 
valid ranges.

Change-Id: I30ac473cca323e4f7bb0ca2137fcb0eee18a806a
Reviewed-on: https://review.monogon.dev/c/monogon/+/3069
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI