m/n/c/network: implement ARP announcements
This implements ARP announcements in the Metropolis network stack.
Its intent is to help IP stacks on the same broadcast domain to update
their ARP entries once a Metropolis network stack comes up.
The format of the ARP packets is chosen to bypass most EVPN ARP
suppression mechanisms to ensure this also works with these systems.

Change-Id: I2db1248f7034ea56930cf6f4a93de598b0f8c7de
Reviewed-on: https://review.monogon.dev/c/monogon/+/3074
Reviewed-by: Leopold Schabel <leo@monogon.tech>
Tested-by: Jenkins CI