m/test/launch/cli: turn targets back into binaries and aliases
We had some kind of hack to get //metropolis:{launch,launch-cluster} to
work as `bazel test` targets, but I see no reason to ever want that.

Having the thing go through a test_suite and a native_test broke passing
SIGINT and caused the emulated environments to continue running in the
background forked off the server when the user exited via Ctrl-C.

For some reason, running in tests also allowed us to do a weird hack in
which we could resolve the TPM config directory as a runfile and list
files there (running via `bazel run` broke that). Let's also fix this to
just use a file list instead.

Change-Id: I3389617272307275e2755e540b233f88ca80f0bd
Reviewed-on: https://review.monogon.dev/c/monogon/+/3105
Tested-by: Jenkins CI
Reviewed-by: Leopold Schabel <leo@monogon.tech>
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>