metropolis/test: create swtpm TPMs at runtime instead of compile time
The generated TPM data is random (it contains generated cryptographic
keys) so we really shouldn't be building it with Bazel.

Instead, let's create it at runtime for e2e tests, and also actually
generate separate TPM data per node with a common issuer for all.

Moving the logic out of //metropolis/node also feels deserved, as this
is all squarely in test territory.

Change-Id: I257ee54c88ede685ba3faf573282b0f9228b10e8
Reviewed-on: https://review.monogon.dev/c/monogon/+/3132
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>