m/n/c/network: refetch link attributes from kernel
Fetch all link attributes from the kernel once an interface is
created/configured to ensure all attributes are present and any
defaults set by the kernel are taken into account.

Only affects the static configuration path, dynamic configuration path
already gets the interface data directly from the kernel.

This is an alternative to CL 3112 which had to be reverted.

Change-Id: I1076fa318cac5f0ecdb47c5b0aac4dfe7ea2cac9
Reviewed-on: https://review.monogon.dev/c/monogon/+/3139
Tested-by: Jenkins CI
Reviewed-by: Serge Bazanski <serge@monogon.tech>