workspace: make query regexes anchored in gopackagesdriver
Before this fix, when using VS Code with the Go extension, the initial 
query of all go packages from the gopackagesdriver with a ./... query 
failed with the error:
root package @@//cloud/bmaas/bmdb/model:sqlc_model is missing

This happened because an unanchored regex was used to query for 
go_library rules, and this query also returned sqlc_go_library rules. 
This patches rules_go to add the missing regex anchors.

It turns out that this error caused gopls to discard the result of the 
initial query of everything, and it then queried individual files as 
they were opened in the editor. This no longer happens after this fix, 
so the editor features are now available immediately when opening a go 
source file.

Change-Id: I03ec50777d1cc543eb1cd767b58f61bb5b5261f5
Reviewed-on: https://review.monogon.dev/c/monogon/+/3187
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>