m/n/c/n/dhcp4c: minor cleanup
The DiscoverBackoff was not Reset initially, which means the first 
discovery started with a backoff of 500 ms (ExponentialBackOff default) 
instead of 1400 ms.

Remove the unused lastBoundTransition field.

If the backoff has no timeout and stateDeadline is not set, 
stateDeadlineExceeded should not be called (the discovering state does 
not define this function). This was actually already the case before, 
because NextBackOff never returns a duration smaller than 10 ms, but 
this change makes that more clear.

Change-Id: I714a562d8901fba69afaf6b779c6db310577fca5
Reviewed-on: https://review.monogon.dev/c/monogon/+/3215
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>