third_party/xfsprogs: update to 6.9.0
The previous version was three years old and we want a version newer
than 5.15 to enable bigtime by default for new filesystems. This gets us
to the current version of xfsprogs again.

There is an additional dependency on urcu which was added since 5.15
which is added in a parent CL of this, it provides core functionality
and cannot be feasibly patched out.

Closes: https://github.com/monogon-dev/monogon/issues/1
Change-Id: I16b93c3799574030548d6460c298e2980b78a5a8
Reviewed-on: https://review.monogon.dev/c/monogon/+/3346
Tested-by: Jenkins CI
Reviewed-by: Jan Schär <jan@monogon.tech>