osbase/gpt: sync device after writing each GPT
This makes it safer to update the GPT. Without the first sync call, it 
could for example happen that both GPTs are partially written to disk 
when the system loses power, resulting in both GPTs being corrupted. The 
second sync call ensures that the GPT update is committed to disk and 
will not be rolled back, and also that both GPTs are valid after the 
call to Write returns.

Change-Id: I50b3cabee4ee8a3162766812e945e129358dd875
Reviewed-on: https://review.monogon.dev/c/monogon/+/3360
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI