m/n/c/update: implement Rollback
Implement a mechanism for manual rollbacks, useful for cases where
rolling forward is not an option or automated rollbacks did not catch an
issue. To ensure that the rollback does not break the machine, the
alternate slot is only tried on next boot and that version needs to set
the slot active before it is permanently activated.

Change-Id: I2fe4dfedcecd5bf7d1bdebdd070e40e817bca7c3
Reviewed-on: https://review.monogon.dev/c/monogon/+/3386
Reviewed-by: Serge Bazanski <serge@monogon.tech>
Tested-by: Jenkins CI