third_party/linux: disable fallback tunnel devs
Introduces a patch for a kernel config option to control the default
behavior for fallback tunnel devices in the kernel.
We want to fully disable them but this can otherwise only be done by
passing kernel commandline arguments which would need to be specified in
a lot of places and can easily be missed.

This will get rid of things like the sit0 interface in the host and
container namespaces.

Change-Id: I2c03c5aa50bc64b527b72c3c022d3e9f2111dd84
Reviewed-on: https://review.monogon.dev/c/monogon/+/3430
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>