m/n/c/curator: maintain consistency between roles and etcd members
When updating the consensus role, both etcd membership and the role need
to be updated. It is possible that the etcd membership change is applied
but the role update fails, resulting in an inconsistency. This change
adds a background process which cleans up this inconsistency by updating
roles to match etcd membership.

This is partially based on previous work by Serge Bazanski, where this
background sync was performed in the opposite direction: etcd membership
is removed if the role is missing. Here, I instead update the role based
on etcd membership. This has the benefit that we finish partially
applied management operations, instead of fighting them.

Co-authored-by: Serge Bazanski <serge@monogon.tech>
Change-Id: I8871b068d1d20c65bcbea5289eafe54676906819
Reviewed-on: https://review.monogon.dev/c/monogon/+/3438
Tested-by: Jenkins CI
Reviewed-by: Serge Bazanski <serge@monogon.tech>