m/n/c/curator: add ID field to nodes in etcd
This makes the ID independent of the public key for nodes stored in
etcd. This is needed to eventually allow node key rotation.

We could just extract the ID from the key without adding an ID field.
But the consistency check between key and value has already caught a bug
once, so it seems worth keeping.

Change-Id: I7ba5904d37d54e93ad6dc7d4b6f0cfac19bc730d
Reviewed-on: https://review.monogon.dev/c/monogon/+/3475
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>