metropolis/vm: delete vm package
This only contains the CRD right now and is not used/implemented
anywhere. This is also the only usage of kube-code-gen which breaks with
the newest version of rules_go. Since we don't plan to use this
in the near future this also gets deleted, because even if we need it
again, we can just revert this change.

Change-Id: I29ab75541957fce6a7dd8414c0df3cfdf90f8ec3
Reviewed-on: https://review.monogon.dev/c/monogon/+/3465
Tested-by: Jenkins CI
Reviewed-by: Serge Bazanski <serge@monogon.tech>