osbase/build/mkimage: replace embedsrc abloader reference with argument
rules_rust things that our abloader target itself is an exec target
because it is included as embedsrc inside mkimage. To prevent this wrong
detection we provide it as runfile like the kernel and rootfs. This is
a preparation for updating rust to the current stable version, as it
requires specific overrides to work correctly with our efi toolchain.

Change-Id: I78de6a15570a81d9f673702ec4e50954d604117d
Reviewed-on: https://review.monogon.dev/c/monogon/+/3598
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>