osbase/fat32: fix corrupted directories
Previously, after writing the contents of a file or directory, any
remaining space in the last cluster was left as is instead of zeroed.
For files this may be acceptable, but directories do not have an
explicit size in bytes. The result is that, if the device was not zeroed
before writing the FAT32, each directory contains some number of
additional garbage entries, with a random name, size, and initial
cluster.

The fix is to ensure that the last cluster of a directory is filled up
with zeroes, following the FAT32 spec (page 24): "initialize all bytes
of that cluster to 0". The change also applies to files, where zeroing
the remainder of the last cluster should not be necessary, but also
doesn't hurt. At most 32 KiB of zeroes (the maximum cluster size) will
be written for each file.

The fsck test is extended to initialize the file with random bytes.
Without the fix, this results in fsck reporting a large number of
problems.

Change-Id: I7729e028c92bfad9b879a4256f3fa4f7af25553a
Reviewed-on: https://review.monogon.dev/c/monogon/+/3607
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI