osbase/fat32: write file creation time
Add file creation times to directory entries. This was already partially
implemented, and there even was a test. But the test was conditional on
the kernel version, and the version comparison was broken. The test
timestamps were also wrong, as the last component of time.Date is
nanoseconds, not milliseconds.

Change-Id: Ic00d6bb27aced918df018c9ff3148d12ecb07753
Reviewed-on: https://review.monogon.dev/c/monogon/+/3608
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>