build/ci: add bazel mod tidy step
Running `bazel mod tidy` on the current main branch causes a change in
build/bazel/rust.MODULE.bazel. If `bazel run //:go -- mod tidy` changes
go.mod, it will also also run `bazel mod tidy`, which means that this
will eventually be noticed in a later unrelated change. But it is better
if this is noticed and fixed immediately, so I added `bazel mod tidy` to
the steps run by CI. It looks like we then don't need `bazel mod deps`
anymore.

I also moved the git status step to the post always block, such that it
is run even if one of the previous steps fails. For example,
`bazel run //:gazelle -- update` can fail if a change to
MODULE.bazel.lock is necessary. This should make it easier to understand
why the pipeline failed, because you can then still see which files have
changed, and the "Unclean working directory" message is printed.
I tested that this works, and that the pipeline still fails if a step
in the always block fails.

Change-Id: I05e3f1a5d9e45ba16058cc6758b0acb106458cae
Reviewed-on: https://review.monogon.dev/c/monogon/+/3609
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>