osbase/event/memory: remove Sync feature
The Sync feature was unused, and I'm not sure using it anywhere is a
good idea. If you really do need reliable delivery of events to multiple
consumers, then you probably also would need some mechansim for a
consumer to restart at the point in the history it has last seen before
it failed. Removing this feature simplifies the implementation and
allows new features to be implemented.

Change-Id: I94731355b7208f91256bc4610cb31a2720ab6c68
Reviewed-on: https://review.monogon.dev/c/monogon/+/3717
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>